containers/sast-cli-tool/requirements.txt
flask>=3.1.2
flask-cors>=5.0.0
flask-limiter>=3.7.0
pyjwt>=2.9.0
psycopg2-binary>=2.9.9
aiohttp>=3.10.5
aiofiles>=24.1.0
pyyaml>=6.0.2
regex>=2024.9.11
tenacity>=9.0.0
pydantic>=2.9.2